webrequest/requirements.txt
Flask-JSONRPC==3.0.1
Flask[async]==3.1.0
python-dotenv==1.0.1
aiohttp==3.10.11
openai==1.47.0
anthropic==0.34.2
httpx==0.27.2
